Enable typescript-eslint/explicit-function-return-type
We should annotate types of in/out of functions
because programms are built with a bunch of functions.

It makes type safe world easy to annotate types of in/out of functions.
and we should express the type signature as documentation of their functions.

However, this rule is pretty tired for expressions because we use function expression
as a callback. Let's disable its case.