Note the project goal about stylistic issues
As I remember, almost user projects for this ruleset uses prettier.
Thus this project finished the role about purely stylistic issues
and we should only focus a coding practise to avoid bad-patterns.

This addition does not mean to drop rules related to stylistic issues.
But I think it's not a top level priority for this.